fix: fail closed on removed factory subcommands and satisfy pkg-file-count
Address blocking PR review feedback by rejecting unknown you factory lookups as unknown-command errors (including save with --from), consolidating factory CLI tests to stay within the package file-count cap, and syncing main for the hierarchical named-factory path test fix.
